refactor(test-plugins): Export SpecTestMutator
